Abstract manhattan logic to container component to ensure update triggered (#112)
* Fixes opentargets/genetics#315. Abstract manhattan logic to container

* Remove logging
